refactor(core): centralize get_pipeline_job into foldrun_app/core/vertex_utils
get_pipeline_job was defined in foldrun_app/models/af2/utils/vertex_utils.py
and cross-imported directly from the AF2 module by OF3 and Boltz-2 tools,
creating an unintended coupling.

Move the function to foldrun_app/core/vertex_utils (shared infrastructure
layer), re-export it from af2/utils/vertex_utils for backward compatibility
so all existing AF2 callers are unaffected, and update OF3 and Boltz-2 to
import from core directly.
