+++ applications/foldrun/foldrun-agent/foldrun_app/models/boltz2/tools/analyze_job.py
@@ -162,7 +162,7 @@ class BOLTZ2JobAnalysisTool(BOLTZ2Tool):
         logger.info(f"Starting BOLTZ2 parallel analysis for job {job_id}")
 
         # Get job from Vertex AI
-        from foldrun_app.models.af2.utils.vertex_utils import get_pipeline_job
+        from foldrun_app.core.vertex_utils import get_pipeline_job
 
         try:
             job = get_pipeline_job(job_id, self.config.project_id, self.config.region)
